Fix: code changes in controller file

+++ backend/controllers/product.controller.js
@@ -52,7 +52,11 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const { id } = req.params
-    console.log("id:", id)
+
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid Product id" })
+    }
 
     try {
         await Product.findByIdAndDelete(id);
@@ -60,6 +64,6 @@ export const deleteProduct = async (req, res) => {
     }
     catch (error) {
         console.log(" ⛔ Error in fetching product", error.message)
-        res.status(404).json({ success: false, message: "Product not found" });
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 }
